chore: rename EndpointsConfig to resolveEndpointsConfig
