+++ packages/config-resolver/src/endpointsConfig/resolveEndpointsConfig.ts
